I lied (It passes all of the tests now...)
updated the createToken() method in AuthService so that it is guaranteed to produce a unique authToken each time (I think before if it was called too close it would maybe make duplicates or something, but now each authToken is guaranteed to be at least +1 more than the last. It's a pretty quick and dirty solution.